Change Main file, Add initial implementation
